+++ bot/__init__.py
@@ -1,3 +1,6 @@
 import telebot
 import spamwatch
+import nudenet
+import os
+import httplib2
 from .load_environ import Environment