+++ bot/load_environ.py
@@ -20,6 +20,8 @@ class Environment:
     def __load_environ(self):
         self.TG_BOT_TOKEN = os.getenv("TG_BOT_TOKEN")
         self.SPAMWATCH_API_KEY = os.getenv("SPAMWATCH_API_KEY")
+        self.OWNER_NAME = os.getenv("OWNER_NAME")
+        self.OWNER_USERNAME = os.getenv("OWNER_USERNAME")
         
     def __validate_bot_token(self):
         if self.TG_BOT_TOKEN is None or len(self.TG_BOT_TOKEN) == 0: